Refactor Filter into common components
